Move MetaBinding and MetaCallMapping to declarations
