+++ OOModel/src/declarations/MetaBinding.cpp
